+++ src/utils.ts
@@ -4,7 +4,7 @@
  * Utilities used by the Mangrove smart contract repos.
  */
 
-import { Erc20, Erc20Id, Role, UniqueId } from "./types";
+import { Erc20, Erc20Id, Erc20Instance, Role, UniqueId } from "./types";
 
 /** Network names used in Mangrove smart contract repos.
  *
@@ -111,20 +111,8 @@ export function toNamedAddressesPerNamedNetwork(
  */
 export function toErc20InstancesPerNamedNetwork(
   erc20s: Record<Erc20Id, Erc20>,
-): Record<
-  string,
-  { symbol: string; id: string; address: string; default: boolean }[]
-> {
-  const instancesPerNamedNetwork: Record<
-    string,
-    {
-      symbol: string;
-      decimals: number;
-      id: string;
-      address: string;
-      default: boolean;
-    }[]
-  > = {};
+): Record<string, Erc20Instance[]> {
+  const instancesPerNamedNetwork: Record<string, Erc20Instance[]> = {};
 
   for (const [, erc20] of Object.entries(erc20s)) {
     for (const [networkId, networkInstances] of Object.entries(