+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Next version
 
+- Fix return type for `toErc20InstancesPerNamedNetwork`: Missing `decimals` field added.
+
 # 1.1.0
 
 - Add utils for transforming query results to structures that can be easily serialized the JSON format used by the Mangrove smart contract repos.