+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Next version
 
+- Fix issue in mapping of default tokens in `toErc20InstancesPerNamedNetwork`: Only the original instance ID should be marked as default, not the copy that has the symbol as ID.
+
 # 1.1.0
 
 - Add utils for transforming query results to structures that can be easily serialized the JSON format used by the Mangrove smart contract repos.