+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Next version
 
+- Fix issue in mapping of default tokens in `toErc20InstancesPerNamedNetwork`: Only the original instance ID should be marked as default, not the copy that has the symbol as ID.
 - Fix return type for `toErc20InstancesPerNamedNetwork`: Missing `decimals` field added.
 
 # 1.1.0