+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Next version
 
+# 1.1.1
+
 - Fix issue in mapping of default tokens in `toErc20InstancesPerNamedNetwork`: Only the original instance ID should be marked as default, not the copy that has the symbol as ID.
 - Fix return type for `toErc20InstancesPerNamedNetwork`: Missing `decimals` field added.